Studio Panels: read URL params via @wordpress/route's useSearch in list stage
The wp-build route loader pattern requires useLoaderData() to read the data
in the stage, but @wordpress/route 0.5.0 (the shipping npm version) doesn't
export that hook. Drop the loader and read URL search params directly with
useSearch() — same effect, fewer indirection layers, works with the route
package version Gutenberg currently ships.

Bumps panels to v0.2.3.
