+++ apps/studio-panels/routes/list/package.json
@@ -16,5 +16,8 @@
 		"@wordpress/dataviews": "^14.2.0",
 		"@wordpress/element": "^6.18.0",
 		"@wordpress/i18n": "^6.18.0"
+	},
+	"devDependencies": {
+		"@wordpress/route": "^0.5.0"
 	}
 }
